+++ lib/elixir/test/elixir/inspect_test.exs
@@ -297,7 +297,7 @@ defmodule Inspect.MapTest do
           "inspecting %{__struct__: Inspect.MapTest.Failing, key: 0}"
 
     assert_raise Inspect.Error, msg, fn ->
-      inspect(%Failing{})
+      inspect(%Failing{}, safe: false)
     end
 
     assert [{Inspect.Inspect.MapTest.Failing, :inspect, 2, _}|_] = System.stacktrace
@@ -308,7 +308,7 @@ defmodule Inspect.MapTest do
           "%{__struct__: Inspect.MapTest.Failing, key: 0}` while " <>
           "inspecting %{__struct__: Inspect.MapTest.Failing, key: 0}"
 
-    assert inspect(%Failing{}, safe: true) ==
+    assert inspect(%Failing{}) ==
            "%Inspect.Error{message: \"#{msg}\"}"
   end